customMQTT/requirements.txt
paho-mqtt>=1.6.1
pika>=1.3.0
flask>=2.0.0
numpy>=1.21.0
pymongo>=4.0.0
